Mouse acceleration (#19)
* Make mouse increase its speed over time

Based on a new setting: mouseAcceleration

* Possible to set mouseStartSpeed

Allows the mouse to have some initial speed even before it starts
accelerating, instead of starting from 0.

* Mouse decelerates instead of stopping immediately

Added setting mouseDeceleration for this.

* Need to multiply MouseStartSpeed by time too

* Add acceleration/deceleration settings to example in readme

* Rename mouseStartSpeed to startMouseSpeed

To be consistent with baseMouseSpeed

* Replace MouseAcceleration → MouseAccelerationTime

It's more intuitive to say how long time it should take to reach the
top speed. Acceleration is calculated from the MouseAccelerationTime
setting.

* Configurable acceleration/deceleration curves

Added settings:

* mouseAccelerationCurve
* mouseDecelerationCurve
* mouseDecelerationTime

The setting mouseDeceleration was removed.

Had to remove the special handling of diagonal movement.
That would be treated like a deceleration which gives unexpected
results.