Fix device reconnects (#89)
* remove buggy readloop

From the perspective of a reconnection mechanism the ReadLoop simply
isn't correct. One cannot assume that device nodes (eg.
/dev/input/eventX) will belong to the same device as it did when
mouseless started and initially found all keyboards (or just the ones
defined in the config).

I ran into this issue with a bluetooth keyboard and mouse.  Depending of
which connected/disconnected first mouseless ended up grabbing my mouse
sometimes, only because it got the device node of the keyboard when it
reconnected after a blutooth power saving cycle.

And from the perspective of a safe mechanism of opening the devices in
case of (race conditions?) or other errors it's not really needed (if
that was the intention?), evdev.ListInputDevices already opens the
devices, there is no need to run the ReadLoop just to open the devices
again.

* refactor keyboard check into function

* add hotplug support for keyboards via fsnotify

This is nice for bluetooth devices that go in powersave mode and keep
dis/reconnecting.

If there are devices specified via config file it will only watch for
those otherwise for all.